+++ changelog.md
@@ -1,6 +1,14 @@
 # coreActivity
 ## Changelog
 
+### Version: 1.2 / October 31 2023
+
+* **new** database: logs table has new `country_code` column
+* **new** logging: options for logging country code and other location information
+* **new** geolocation settings: choose between online and IP2Location database
+* **edit** Dev4Press Library 4.4 Beta
+* **fix** logs override filtering not working properly always
+
 ### Version: 1.1 / October 16 2023
 
 * **new** component: WooCommerce plugin, with 3 events